tu-proyecto/backend/requirements.txt
fastapi
uvicorn
requests
trafilatura
PyPDF2
python-dotenv
openai